refactor sanitizer

+++ src/controller/router/sum-router/validator.ts
@@ -2,6 +2,10 @@ import { check } from 'express-validator';
 import validateRequest from '../../helper/validate-request';
 
 export const sumValidator = validateRequest([
-  check('a', 'first number is invalid').isNumeric(),
-  check('b', 'second number is invalid').isNumeric(),
+  check('a', 'first number is invalid')
+    .isNumeric()
+    .toInt(),
+  check('b', 'second number is invalid')
+    .isNumeric()
+    .toInt(),
 ]);
